+++ release-notes-3.3.0.md
@@ -8,5 +8,6 @@ ${version-number}
   - Secrets can be retrieved from the following sources:
     - Environment variables (direct value) - enabled via `CAF_ENABLE_ENV_SECRETS` (defaults to `true`)
     - File content (path specified by environment variable with `_FILE` suffix) - enabled via `CAF_ENABLE_FILE_SECRETS` (defaults to `false`)
+- US975260: Refactor to remove HPE references in package names
 
 #### Known Issues
