+++ staging-service-internal-client/pom.xml
@@ -23,7 +23,7 @@
 
     <parent>
         <artifactId>staging-service-aggregator</artifactId>
-        <groupId>com.github.cafdataprocessing</groupId>
+        <groupId>com.github.cafdataprocessing.services.staging</groupId>
         <version>3.3.0-SNAPSHOT</version>
     </parent>
