+++ staging-service/src/main/java/com/github/cafdataprocessing/services/staging/dao/filesystem/SubBatchWriter.java
@@ -20,8 +20,8 @@ import com.github.cafdataprocessing.services.staging.dao.InputStreamSupplier;
 import com.github.cafdataprocessing.services.staging.exceptions.IncompleteBatchException;
 import com.github.cafdataprocessing.services.staging.exceptions.InvalidBatchException;
 import com.github.cafdataprocessing.services.staging.exceptions.StagingException;
+import com.github.cafdataprocessing.workers.document.schema.validator.InvalidDocumentException;
 import com.github.cafdataprocessing.services.staging.utils.JsonMinifier;
-import com.microfocus.caf.worker.document.schema.validator.InvalidDocumentException;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileNotFoundException;
