+++ staging-service/src/main/java/com/github/cafdataprocessing/services/staging/utils/JsonMinifier.java
@@ -21,8 +21,8 @@ import com.fasterxml.jackson.core.JsonGenerator.Feature;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
 import com.github.cafdataprocessing.services.staging.exceptions.InvalidBatchException;
-import com.microfocus.caf.worker.document.schema.validator.DocumentValidator;
-import com.microfocus.caf.worker.document.schema.validator.InvalidDocumentException;
+import com.github.cafdataprocessing.workers.document.schema.validator.DocumentValidator;
+import com.github.cafdataprocessing.workers.document.schema.validator.InvalidDocumentException;
 import com.worldturner.medeia.api.ValidationFailedException;
 import java.io.IOException;
 import java.io.InputStream;
