+++ worker-batch-ingestion-container-testing/src/test/java/com/github/cafdataprocessing/workers/ingestion/BatchWorkerAcceptanceIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafdataprocessing.worker.ingestion;
+package com.github.cafdataprocessing.workers.ingestion;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
