+++ worker-batch-ingestion-container/worker.sh
@@ -77,6 +77,6 @@ cd /maven
 exec java $CAF_WORKER_JAVA_OPTS \
     -Dpolyglot.engine.WarnInterpreterOnly=false \
     -cp "*:/mnt/mesos/sandbox/*:/mnt/mesos/sandbox/batch-plugins/*" \
-    com.hpe.caf.worker.core.WorkerApplication \
+    com.github.workerframework.core.WorkerApplication \
     server \
     ${dropwizardConfig}
