+++ worker-batch-ingestion-plugin/src/main/java/com/github/cafdataprocessing/workers/ingestion/models/Subbatch.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafdataprocessing.worker.ingestion.models;
+package com.github.cafdataprocessing.workers.ingestion.models;
 
 import com.github.cafdataprocessing.services.staging.BatchId;
 import com.github.cafdataprocessing.services.staging.TenantId;
