+++ worker-batch-ingestion-plugin/src/main/java/com/github/cafdataprocessing/workers/ingestion/validator/FieldValidatorInterface.java
@@ -13,23 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafdataprocessing.worker.ingestion.validator;
+package com.github.cafdataprocessing.workers.ingestion.validator;
 
-import com.hpe.caf.worker.document.DocumentWorkerDocument;
+import com.github.cafdataprocessing.workers.document.DocumentWorkerDocument;
 
-public enum NullFieldValidator implements FieldValidatorInterface
+public interface FieldValidatorInterface
 {
-    INSTANCE;
-
-    @Override
-    public DocumentWorkerDocument validate(final DocumentWorkerDocument document)
-    {
-        return document;
-    }
-
-    @Override
-    public String toString()
-    {
-        return "<null field validator>";
-    }
+    DocumentWorkerDocument validate(DocumentWorkerDocument document);
 }
