+++ worker-batch-ingestion-plugin/src/main/java/com/github/cafdataprocessing/workers/ingestion/validator/ValidationFileAdapter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafdataprocessing.worker.ingestion.validator;
+package com.github.cafdataprocessing.workers.ingestion.validator;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
