--- worker-batch-ingestion-plugin/src/main/resources/META-INF/services/com.hpe.caf.worker.batch.BatchWorkerPlugin
@@ -1,1 +1,0 @@
-com.github.cafdataprocessing.worker.ingestion.IngestionBatchWorkerPlugin
