+++ worker-batch-ingestion-plugin/src/test/java/com/github/cafdataprocessing/workers/ingestion/validator/ValidationFileAdapterTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafdataprocessing.worker.ingestion.validator;
+package com.github.cafdataprocessing.workers.ingestion.validator;
 
 import java.io.IOException;
 import java.util.List;
